Change behavior of fixed exposure feature
The SET button does not have the HOLD property, so we cannot detect when
is released, and thus the fixed exposure feature got stuck. Now, it can
be pressed once to enable the feature, and once again to disable it;
also, when the camera stops metering, the feature is disabled.

Fixes issue #424.